Fix memory leak in overlapping relocation test functions (#7184)
The setup<T>() helper allocates two buffers (mem1, mem2), but overlapping
test blocks only use one buffer for in-place relocation. The second buffer
was captured as a discarded binding (___) and never freed, leaking
N * sizeof(T) bytes per overlapping test block.

This commit introduces a setup_single<T>() helper that allocates only one
buffer, and updates all overlapping test blocks across 6 files to use it:

- uninitialized_relocate.cpp (4 sites)
- uninitialized_relocaten.cpp (4 sites)
- uninitialized_relocate_backward.cpp (3 sites)
- uninitialized_relocate_sender.cpp (3 sites)
- uninitialized_relocaten_sender.cpp (3 sites)
- uninitialized_relocate_backward_sender.cpp (3 sites)

Closes #7184
